+++ src/hooks/useCookie.jsx
@@ -1,18 +1,5 @@
 import { useState } from 'react';
-
-const getItem = (key) => {
-  const cookies = document.cookie.split('; ');
-  return cookies.reduce((total, currentCookie) => {
-    const [storedKey, storedValue] = currentCookie.split('=');
-    return key === storedKey ? decodeURIComponent(storedValue) : total;
-  }, '');
-};
-
-const setItem = (key, value, numberOfDays) => {
-  const now = new Date();
-  now.setTime(now.getTime() + numberOfDays * 60 * 60 * 24 * 1000);
-  document.cookie = `${key}=${value}; expires=${now.toUTCString()}; path=/`;
-};
+import { getItem, setItem } from '../utils/cookies';
 
 export default function useCookie(key, defaultValue) {
   const getCookie = () => getItem(key) || defaultValue;
